+++ spec/dummy/app/models/article.rb
@@ -1,4 +1,4 @@
 class Article < ApplicationRecord
   include RailsAuditLog::Auditable
-  audit_log only: [:title]
+  audit_log only: [:title], meta: { model: -> { "Article" }, title_length: ->(record) { record.title.length } }
 end